introduce new skill templates and refactor structure
