+++ apps/web/lib/skill-templates/prompt.ts
@@ -1,0 +1,36 @@
+import type { SkillTemplateDefinition } from "./types";
+
+function bulletList(items: readonly string[]): string {
+  return items.map((item) => `- ${item}`).join("\n");
+}
+
+export function buildGtmSkillPrompt(template: SkillTemplateDefinition): string {
+  const triggerModes = template.triggerModes
+    .map((mode) => (mode === "scheduled" ? "cron/scheduled agent message" : "manual trigger"))
+    .join(" and ");
+
+  return `I want to create a reusable DenchClaw skill called "${template.title}".
+
+This should become a durable GTM skill, not a one-off chat. DenchClaw is my AI workspace with CRM, Gmail, Calendar, enrichment, web search, and optional HubSpot or Notion context. Treat Dench CRM as the default system of record, and use Gmail and Calendar as first-class context when they are connected.
+
+The desired outcome is:
+${template.outcome}
+
+Available trigger modes for this product are only manual trigger and cron/scheduled agent messages. For this skill, design around ${triggerModes}. Do not assume webhooks, event listeners, or automatic app callbacks exist.
+
+Start by interviewing me one question at a time. Do not echo this setup back to me. Ask the smallest next question needed, keep each interview turn under 120 words unless options are essential, wait for my answer, then ask the next question. Do not create or edit any files until you have enough context to tailor the workflow.
+
+Before writing the final SKILL.md, gather these specifics:
+${bulletList(template.interviewTopics)}
+
+When you have enough context, create a complete SKILL.md for this workflow. Do not assume a skill-creator helper exists; if no helper is available, create the skill directly at skills/<kebab-case-skill-name>/SKILL.md and add references under that folder when useful. The skill should include:
+${bulletList(template.skillInstructions)}
+
+Automation policy:
+- If the skill sends email or LinkedIn-style outreach, it may be fully automated only after you define explicit send rules with me.
+- Capture daily or weekly caps, allowlists or exclusions, quiet hours, stop conditions, duplicate prevention, and what counts as a reply or conversion.
+- For scheduled skills, include idempotency checks so a cron run does not repeat work already done in a previous run.
+- For CRM writes, prefer additive updates with source attribution and confidence notes. Ask before overwriting existing user-authored fields unless I explicitly allow overwrites.
+
+End by showing me the created skill path, how to invoke it manually, and the exact cron/scheduled message to use if the workflow should run on a schedule.`;
+}
